+++ api/config/secret.js
@@ -1,0 +1,5 @@
+const jwtSecret = process.env.SECRET || 'water my plants'
+
+module.exports = {
+    jwtSecret
+}
