--- config/secret.js
@@ -1,5 +1,0 @@
-const jwtSecret = process.env.SECRET || 'water my plants'
-
-module.exports = {
-    jwtSecret
-}
